masscan_nmap/requirements.txt
python-nmap==0.7.1
xmltodict==0.13.0